Add CSS for better rendering in print media type
Currently, the rendered Markdown looks fine on screen, but does not
print well: the "back to top" link leaves a white box on the page that
overlays some of the text, and some of the text on the right margin is
cut off.

This commit tweaks the CSS (conditional on a print media type, so that
the on-screen rendering is not changed) to reset some of the properties.
With this change, the output looks much better in print.